refactor[#64]: dto 위치 수정
