+++ src/main/java/org/highfive/backend/content/controller/OnboardingController.java
@@ -3,11 +3,9 @@ package org.highfive.backend.content.controller;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.highfive.backend.content.dto.OnboardingInitContentsResponseDto;
-import org.highfive.backend.content.dto.mapper.ContentMapper;
+import org.highfive.backend.content.dto.response.OnboardingInitContentsResponseDto;
 import org.highfive.backend.content.dto.request.OnboardingSelectContentRequestDto;
 import org.highfive.backend.content.dto.response.OnboardingSelectContentResponseDto;
-import org.highfive.backend.content.entity.Content;
 import org.highfive.backend.content.service.ContentService;
 import org.highfive.backend.content.service.OnboardingService;
 import org.highfive.backend.global.code.SuccessCode;
